[FEATURE] Introduce ParserRuntimeOnly trait
Implement in ViewHelpers which only have a purpose
during parse time - such as f:comment, f:section
and f:layout.
