+++ src/ViewHelpers/CommentViewHelper.php
@@ -9,6 +9,7 @@ namespace TYPO3Fluid\Fluid\ViewHelpers;
 use TYPO3Fluid\Fluid\Core\Compiler\TemplateCompiler;
 use TYPO3Fluid\Fluid\Core\Parser\SyntaxTree\ViewHelperNode;
 use TYPO3Fluid\Fluid\Core\ViewHelper\AbstractViewHelper;
+use TYPO3Fluid\Fluid\Core\ViewHelper\Traits\ParserRuntimeOnly;
 
 /**
  * This ViewHelper prevents rendering of any content inside the tag
@@ -47,6 +48,7 @@ use TYPO3Fluid\Fluid\Core\ViewHelper\AbstractViewHelper;
  */
 class CommentViewHelper extends AbstractViewHelper
 {
+    use ParserRuntimeOnly;
 
     /**
      * @var boolean
@@ -57,27 +59,4 @@ class CommentViewHelper extends AbstractViewHelper
      * @var boolean
      */
     protected $escapeOutput = false;
-
-    /**
-     * Comments out the tag content
-     *
-     * @return string
-     * @api
-     */
-    public function render()
-    {
-    }
-
-    /**
-     * @param string $argumentsName
-     * @param string $closureName
-     * @param string $initializationPhpCode
-     * @param ViewHelperNode $node
-     * @param TemplateCompiler $compiler
-     * @return null
-     */
-    public function compile($argumentsName, $closureName, &$initializationPhpCode, ViewHelperNode $node, TemplateCompiler $compiler)
-    {
-        return null;
-    }
 }
